Shop buying

+++ Assets/Scripts/Data/WeaponDataSO.cs
@@ -7,7 +7,7 @@ public class WeaponDataSO : ScriptableObject
 {
     [field: SerializeField] public string WeaponName{get;set;}
     [field: SerializeField] public Sprite Sprite{get;set;}
-    [field: SerializeField] public float price{get;set;}
+    [field: SerializeField] public int price{get;set;}
     [field: SerializeField] public Weapon weapon{get;set;}
     [HorizontalLine]
     [SerializeField] private float Attack;
